requirements.txt
liquidctl>=1.16.0
pystray>=0.19.0
Pillow>=10.0.0
